image-service/requirements.txt
fastapi
uvicorn[standard]
torch
transformers
accelerate
Pillow
ultralytics